backend/requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
langchain
langchain-google-genai
langchain-community
sentence-transformers
torch 